+++ cmd/root_test.go
@@ -1,9 +1,10 @@
 package cmd
 
 import (
-	"github.com/spf13/viper"
 	"os"
 	"testing"
+
+	"github.com/spf13/viper"
 )
 
 // Test the stub root command
