+++ cmd/status.go
@@ -4,20 +4,21 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"os"
+
 	"github.com/algorandfoundation/hack-tui/internal"
 	"github.com/algorandfoundation/hack-tui/ui"
 	"github.com/algorandfoundation/hack-tui/ui/style"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"os"
 )
 
 // statusCmd is the main entrypoint for the `status` cobra.Command with a tea.Program
 var statusCmd = &cobra.Command{
 	Use:   "status",
 	Short: "Get the node status",
-	Long:  style.Purple(BANNER) + "\n" + style.LightBlue("View the node status"),
+	Long:  style.Purple(style.BANNER) + "\n" + style.LightBlue("View the node status"),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		initConfig()
 		if viper.GetString("server") == "" {
